requirements.txt
bcrypt==2.0.0
cffi==1.6.0
Flask==0.10.1
Flask-Bcrypt==0.7.1
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.1
gunicorn==19.5.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mysqlclient==1.3.7
psycopg2==2.6.1
pycparser==2.14
six==1.10.0
SQLAlchemy==1.0.13
Werkzeug==0.11.9
